+++ static/audio.html
@@ -6,8 +6,7 @@
     <title>Audio System</title>
     <meta name='viewport' content='width=device-width, initial-scale=1'>
     <script src="/socket.io/socket.io.js"></script>
-    <script type="module" src="/socketBase.js"></script>
-    <script type="module" src="/audio.js"></script>
+    <script type="module" src="/js/audio.js"></script>
 </head>
 <body>
 </body>
