+++ static/control.html
@@ -6,10 +6,10 @@
     <meta http-equiv='X-UA-Compatible' content='IE=edge'>
     <title>Control Page</title>
     <meta name='viewport' content='width=device-width, initial-scale=1'>
-    <link rel='stylesheet' type='text/css' media='screen' href='control.css'>
+    <link rel='stylesheet' type='text/css' media='screen' href='/css/control.css'>
     <script src="/socket.io/socket.io.js"></script>
-    <script type="module" src='control.js'></script>
-    <script src='qrcode.min.js'></script>
+    <script type="module" src='/js/control.js'></script>
+    <script src='/js/qrcode.min.js'></script>
 </head>
 
 <body>
