+++ static/css/game.css
@@ -1,4 +1,4 @@
-@import url("/common.css");
+@import url("/css/common.css");
 
 body {
     background-color: lime;
