+++ static/css/schedule.css
@@ -1,4 +1,4 @@
-@import url("/common.css");
+@import url("/css/common.css");
 
 body {
     font-size: 4vh;
