+++ static/game.html
@@ -5,11 +5,9 @@
     <meta http-equiv='X-UA-Compatible' content='IE=edge'>
     <title>Game View</title>
     <meta name='viewport' content='width=device-width, initial-scale=1'>
-    <link rel='stylesheet' type='text/css' media='screen' href='game.css'>
+    <link rel='stylesheet' type='text/css' media='screen' href='/css/game.css'>
     <script src="/socket.io/socket.io.js"></script>
-    <script type="module" src="/socketBase.js"></script>
-    <script type="module" src="/gamebase.js"></script>
-    <script type="module" src='game.js'></script>
+    <script type="module" src='/js/game.js'></script>
 </head>
 <body>
     <section id="scoreboard">
