+++ static/input.html
@@ -5,9 +5,9 @@
     <meta http-equiv='X-UA-Compatible' content='IE=edge'>
     <title>Input Page</title>
     <meta name='viewport' content='width=device-width, initial-scale=1'>
-    <link rel='stylesheet' type='text/css' media='screen' href='input.css'>
-    <script src='/socket.io/socket.io.js'></script>
-    <script type="module" src='input.js'></script>
+    <link rel='stylesheet' type='text/css' media='screen' href='/css/input.css'>
+    <script src="/socket.io/socket.io.js"></script>
+    <script type="module" src='/js/input.js'></script>
 </head>
 <body>
     <section>
