+++ static/js/scoreboard.js
@@ -1,4 +1,4 @@
-import {socket, init} from "/socketbase.js";
+import {socket, init} from "/js/socketbase.js";
 
 init(["matchSaved"], ()=>{
     socket.emit("getScoreboard");
