+++ static/js/socketbase.js
