+++ static/schedule.html
@@ -5,10 +5,9 @@
     <meta http-equiv='X-UA-Compatible' content='IE=edge'>
     <title>Input Page</title>
     <meta name='viewport' content='width=device-width, initial-scale=1'>
-    <link rel='stylesheet' type='text/css' media='screen' href='schedule.css'>
+    <link rel='stylesheet' type='text/css' media='screen' href='/css/schedule.css'>
     <script src="/socket.io/socket.io.js"></script>
-    <script type="module" src='socketbase.js'></script>
-    <script type="module" src='schedule.js'></script>
+    <script type="module" src='/js/schedule.js'></script>
 </head>
 <body>
     <div class="match" style="margin:1em auto">
