+++ static/scoreboard.html
@@ -5,10 +5,9 @@
     <meta http-equiv='X-UA-Compatible' content='IE=edge'>
     <title>Input Page</title>
     <meta name='viewport' content='width=device-width, initial-scale=1'>
-    <link rel='stylesheet' type='text/css' media='screen' href='scoreboard.css'>
+    <link rel='stylesheet' type='text/css' media='screen' href='/css/scoreboard.css'>
     <script src="/socket.io/socket.io.js"></script>
-    <script type='module' src='socketbase.js'></script>
-    <script type='module' src='scoreboard.js'></script>
+    <script type='module' src='/js/scoreboard.js'></script>
 </head>
 <body id="body">
     <div class="layer" id="layer-0" style="opacity: 100%;">
